+++ packages/terra-collapsible-menu-view/tests/nightwatch/FirstHiddenItemIsIcon.jsx
@@ -14,7 +14,6 @@ const CollapsibleMenuViewDemo = () => (
       text="Button3"
       className="IconOnly"
       key="button3"
-      isIconOnly
     />
   </CollapsibleMenuView>
 );