+++ packages/terra-collapsible-menu-view/tests/wdio/collapsible-menu-view-spec.js
@@ -5,8 +5,6 @@ const ignoredA11y = {
   'button-name': { enabled: false },
   // https://github.com/cerner/terra-core/issues/1061
   'aria-allowed-attr': { enabled: false },
-  // https://github.com/cerner/terra-core/issues/1066
-  list: { enabled: false },
 };
 
 describe('Collapsible Menu View', () => {