+++ packages/terra-site/src/Index.jsx
@@ -24,6 +24,7 @@ import EmbeddedContentConsumerExamples from './examples/embedded-content-consume
 import FormExamples from './examples/form/Index';
 import FormCheckboxExamples from './examples/form-checkbox/Index';
 import FormFieldExamples from './examples/form-field/Index';
+import FormRadioExamples from './examples/form-radio/Index';
 import FormTextareaExamples from './examples/form-textarea/Index';
 import GridExamples from './examples/grid/Index';
 import HeadingExamples from './examples/heading/Index';
@@ -47,6 +48,7 @@ import SlideGroupExamples from './examples/slide-group/Index';
 import SlidePanelExamples from './examples/slide-panel/Index';
 import SpacerExamples from './examples/spacer/Index';
 import StatusExamples from './examples/status/Index';
+import StatusViewExamples from './examples/status-view/Index';
 import TableExamples from './examples/table/Index';
 import TabsExamples from './examples/tabs/Index';
 import TextExamples from './examples/text/Index';
@@ -76,6 +78,7 @@ import FormTestRoutes from 'terra-form/tests/nightwatch/FormTestRoutes';
 import FormCheckboxTestRoutes from 'terra-form-checkbox/tests/nightwatch/CheckboxTestRoutes';
 import FormFieldTestRoutes from 'terra-form-field/tests/nightwatch/FieldTestRoutes';
 import FormIndexRoutes from './examples/form/FormIndexRoutes';
+import FormRadioTestRoutes from 'terra-form-radio/tests/nightwatch/RadioTestRoutes';
 import FormTextareaRoutes from 'terra-form-textarea/tests/nightwatch/TextareaTestRoutes';
 import GridTestRoutes from 'terra-grid/tests/nightwatch/GridTestRoutes';
 import HeadingTestRoutes from 'terra-heading/tests/nightwatch/HeadingTestRoutes';
@@ -115,6 +118,7 @@ import SlidePanelTestRoutes from 'terra-slide-panel/tests/nightwatch/SlidePanelT
 import SpacerTestRoutes from 'terra-spacer/tests/nightwatch/SpacerTestRoutes';
 import SingleSelectListTestRoutes from 'terra-list/tests/nightwatch/single-select-list/SingleSelectListTestRoutes';
 import StatusTestRoutes from 'terra-status/tests/nightwatch/StatusTestRoutes';
+import StatusViewTestRoutes from 'terra-status-view/tests/nightwatch/StatusViewTestRoutes';
 import TableIndexRoutes from './examples/table/TableIndexRoutes';
 import TableTestRoutes from 'terra-table/tests/nightwatch/table/TableTestRoutes';
 import TabsTestRoutes from 'terra-tabs/tests/nightwatch/Tabs/TabsTestRoutes';
@@ -152,6 +156,7 @@ ReactDOM.render((
       <Route path="form-checkbox" component={FormCheckboxExamples} />
       <Route path="form-textarea" component={FormTextareaExamples} />
       <Route path="form-field" component={FormFieldExamples} />
+      <Route path="form-radio" component={FormRadioExamples} />
       <Route path="grid" component={GridExamples} />
       <Route path="heading" component={HeadingExamples} />
       <Route path="hookshot" component={HookshotExamples} />
@@ -174,6 +179,7 @@ ReactDOM.render((
       <Route path="slide-panel" component={SlidePanelExamples} />
       <Route path="spacer" component={SpacerExamples} />
       <Route path="status" component={StatusExamples} />
+      <Route path="status-view" component={StatusViewExamples} />
       <Route path="table" component={TableExamples} />
       <Route path="tabs" component={TabsExamples} />
       <Route path="text" component={TextExamples} />
@@ -234,6 +240,7 @@ ReactDOM.render((
     {FormCheckboxTestRoutes}
     {FormTextareaRoutes}
     {FormFieldTestRoutes}
+    {FormRadioTestRoutes}
     {OverlayTestRoutes}
     {LoadingOverlayTestRoutes}
     {OverlayContainerTestRoutes}
@@ -248,6 +255,7 @@ ReactDOM.render((
     {HookshotTestRoutes}
     {SelectTestRoutes}
     {SelectOptionTestRoutes}
+    {StatusViewTestRoutes}
     {TabsTestRoutes}
     {TabPaneTestRoutes}
   </Router>