+++ packages/terra-site/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+1.22.0 - (January 18, 2018)
+------------------
 ### Changed
 * Add context to selectable list example